[refactor] implements 로 변경
