+++ src/main/java/bowling/domain/Frame.java
@@ -1,47 +1,20 @@
 package bowling.domain;
 
-import java.util.ArrayList;
 import java.util.List;
 
-public abstract class Frame {
+public interface Frame {
 
-    public static final int FRAME_MAX_SCORE = 10;
-    public static final int SPARE_CHECK_SIZE = 2;
+    void addScore(int score);
 
-    public List<Integer> scores = new ArrayList<>();
-    public boolean isEnd;
-    public boolean isSpare;
+    boolean isEndCondition(int score);
 
+    void checkIsSpare();
 
-    public void addScore(int score) {
-        this.scores.add(score);
-        this.isEnd = isEndCondition(score);
-        checkIsSpare();
-    }
+    boolean hasSize(int size);
 
-    public abstract boolean isEndCondition(int score);
+    List<Integer> getScores();
 
-    public void checkIsSpare() {
-        if (scores.size() != SPARE_CHECK_SIZE) {
-            return;
-        }
+    boolean isSpare();
 
-        isSpare = scores.get(0) + scores.get(1) == FRAME_MAX_SCORE;
-    }
-
-    public boolean hasSize(int size) {
-        return scores.size() == size;
-    }
-
-    public List<Integer> getScores() {
-        return scores;
-    }
-
-    public boolean isSpare() {
-        return isSpare;
-    }
-
-    public boolean isEnd() {
-        return isEnd;
-    }
+    boolean isEnd();
 }
