+++ src/main/java/bowling/domain/GameRecord.java
@@ -53,4 +53,5 @@ public class GameRecord {
     public boolean isEnd() {
         return isEnd;
     }
+
 }
