+++ src/test/java/bowling/domain/NormalFrameTest.java
@@ -10,7 +10,7 @@ class NormalFrameTest {
     @Test
     @DisplayName("스트라이크 치면 1Frame 종료된다.")
     void isEndByStrike() {
-        NormalFrame frame = new NormalFrame(10);
+        NormalFrame frame = new NormalFrame(1,10);
 
         assertThat(frame.isEnd()).isTrue();
     }
@@ -18,7 +18,7 @@ class NormalFrameTest {
     @Test
     @DisplayName("스트라이크가 아닌 투구는 종료되지 않는다.")
     void isEndByTwoStep() {
-        NormalFrame frame = new NormalFrame(8);
+        NormalFrame frame = new NormalFrame(1,8);
 
         assertThat(frame.isEnd()).isFalse();
     }
